requirements.txt
Flask==2.2.5
python-dotenv==1.0.0
requests==2.31.0
google-cloud-vision
spotipy==2.23.0
